grpc client compatible with apitest
